+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * feat(ProseCard): new component
 * feat(ProseCardGroup): new component
 * feat(ProseCodeGroup): new component
+* feat(ProseCodePreview): new component
 * feat(Banner): new component
 * feat(Card): new component
 * feat(CheckboxGroup): new component